Major refactor of config tested and working

+++ scripts/utils/config.py
@@ -1,44 +1,59 @@
 """Runtime configuration for the workflow."""
 
+import os
 import csv
 import json
 from Bio import SeqIO
-from .env import getenv
-
-CANDIDATE_CSV_HEADER = [
-    "species",
-    "taxid",
-    "accession",
-    "hit_subject",  # rename
-    "identity",  # rename, recalc
-    "query_coverage",    # rename, recalc
-    "alignment_length",  # TODO
-    "e_value",
-    "bitscore",  # rename
-]
+from pathlib import Path
+from .flags import FLAG_DETAILS
 
 
 class Config:
 
+    output_dir = Path(os.getenv("OUTPUT_DIR", 'output'))
+    INPUT_TOI_FILEPATH = Path(
+        os.getenv(
+            "INPUT_TOI_FILEPATH",
+            'tests/test-data/taxa_of_interest.txt')
+    )
+    ACCESSIONS_FILENAME = os.getenv("ACCESSIONS_FILENAME", "accessions.txt")
+    TAXONOMY_FILE = os.getenv("TAXONOMY_FILENAME", 'taxonomy.csv')
+    QUERY_TITLE_FILE = os.getenv("QUERY_TITLE_FILENAME", 'query_title.txt')
+    HITS_JSON = os.getenv("HITS_JSON_FILENAME", 'hits.json')
+    HITS_FASTA = os.getenv("HITS_FASTA_FILENAME", 'hits.fasta')
+    FLAGS_CSV = os.getenv("FLAGS_CSV_FILENAME", 'flags.csv')
+    TAXONOMY_ID_FILE = os.getenv("TAXONOMY_ID_FILENAME",
+                                 'assigned_taxonomy.txt')
+    CANDIDATES_FASTA = os.getenv("CANDIDATES_FASTA_FILENAME",
+                                 'candidates.fasta')
+    CANDIDATES_CSV = os.getenv("CANDIDATES_CSV_FILENAME", 'candidates.csv')
+    CANDIDATES_JSON = os.getenv("CANDIDATES_JSON_FILENAME", 'candidates.json')
+    TOI_DETECTED_CSV = os.getenv("TOI_DETECTED_CSV_FILENAME",
+                                 'taxa_of_concern_detected.csv')
+
+    class ALIGNMENT:
+        MIN_NT = int(os.getenv('MIN_NT', 400))
+        MIN_Q_COVERAGE = float(os.getenv('MIN_Q_COVERAGE', 0.85))
+        MIN_IDENTITY = float(os.getenv('MIN_IDENTITY', 0.935))
+        MIN_IDENTITY_STRICT = float(os.getenv('MIN_IDENTITY_STRICT', 0.985))
+
     class FLAGS:
         """Flags for reporting outcomes."""
+        CANDIDATES = 1
+        TOI = 2
         A = "A"
         B = "B"
         C = "C"
         D = "D"
         E = "E"
 
-    def __init__(self, output_dir):
-        self.output_dir = output_dir
-        self.INPUT_TOI_FILE = getenv("INPUT_TOI_FILE")
-        self.HITS_JSON = getenv("HITS_JSON")
-        self.HITS_FASTA = getenv("HITS_FASTA")
-        self.TAXONOMY_FILE = output_dir / getenv("TAXONOMY")
-        self.FLAGS_CSV = getenv("FLAGS_CSV")
-        self.ID_FILE = getenv("ID_FILE")
-        self.CANDIDATES_FASTA = getenv("CANDIDATES_FASTA")
-        self.CANDIDATES_CSV = getenv("CANDIDATES_CSV")
-        self.CANDIDATES_JSON = getenv("CANDIDATES_JSON")
+    @property
+    def taxonomy_path(self):
+        return self.output_dir / self.TAXONOMY_FILE
+
+    def set_output_dir(self, output_dir):
+        self.output_dir = Path(output_dir)
+        self.output_dir.mkdir(exist_ok=True, parents=True)
 
     def read_blast_hits_json(self, query_dir):
         """Read BLAST hits from JSON file."""
@@ -50,59 +65,46 @@ class Config:
         path = query_dir / self.HITS_FASTA
         return self._read_fasta(path)
 
-    def read_toi_file(self):
+    def read_taxa_of_interest(self) -> list[str]:
         """Read taxa of interest from TOI file."""
         return [
             line.strip()
-            for line in self.INPUT_TOI_FILE.read_text().splitlines()
+            for line in self.INPUT_TOI_FILEPATH.read_text().splitlines()
             if line.strip()
         ]
 
     def read_taxonomy_file(self):
         """Read taxonomy from CSV file."""
         taxonomies = {}
-        with self.TAXONOMY_FILE.open() as f:
+        with self.taxonomy_path.open() as f:
             for row in csv.DictReader(f):
-                taxonomies[row["accession"]] = row
+                taxonomies[row["accession"].split('.')[0]] = row
         return taxonomies
 
-    def write_candidates(self, query_ix, hits, species):
-        """Write candidate hits and species to CSV file."""
-        self._write_candidates_json(query_ix, hits, species)
-        self._write_candidates_csv(query_ix, hits, species)
-        self._write_candidates_fasta(query_ix, species)
-
-    def _write_candidates_json(self, query_ix, hits, species):
-        path = (
-            self.output_dir
-            / self.get_query_dir(query_ix)
-            / self.CANDIDATES_JSON)
-        with path.open("w") as f:
-            json.dump({
-                "hits": hits,
-                "species": species,
-            }, f)
-
-    def _write_candidates_csv(self, query_ix, species):
-        path = (
-            self.output_dir
-            / self.get_query_dir(query_ix)
-            / self.CANDIDATES_CSV)
-        with path.open("w") as f:
+    def write_flag(self, query_ix, flag_num, outcome):
+        """Write flags to CSV file."""
+        path = self.get_query_dir(query_ix) / self.FLAGS_CSV
+        if not path.exists():
+            with path.open("w") as f:
+                writer = csv.writer(f)
+                writer.writerow(["name", "number", "outcome", "explanation"])
+        with path.open("a") as f:
             writer = csv.writer(f)
-            writer.writerow(CANDIDATE_CSV_HEADER)
-            for hit in species:
-                writer.writerow([
-                    hit.get(key, "")
-                    for key in CANDIDATE_CSV_HEADER
-                ])
-
-    def _write_candidates_fasta(query_ix, species):
-        """Write FASTA sequences for each candidate species to file."""
-        pass
+            writer.writerow([
+                FLAG_DETAILS[flag_num]['name'],
+                flag_num,
+                outcome,
+                FLAG_DETAILS[flag_num]['explanation'][outcome],
+            ])
 
     def get_query_dir(self, query_ix):
-        return self.output_dir / f"query_{query_ix}"
+        d = self.output_dir / f"query_{query_ix}"
+        d.mkdir(exist_ok=True, parents=True)
+        return d
+
+    def ix_from_query_dir(self, query_dir):
+        query_dir = Path(query_dir)
+        return int(query_dir.name.split("_")[1])
 
     def _read_json(self, path):
         """Read JSON file."""
